Allow customer headers during JWT generation (#3)
This allows custom headers to be included in the options passed to
the to_jwt method, so that e.g. key id and type headers can be included.
This also adds a convenience constant for the JWT type commonly used in
the type header. Lastly, it calls #deep_normalize_hash on the
additional_claims hash to allow indifferent access.